Split pdbe.io module into io and structure modules
