+++ src/protein_quest/mcp_server.py
@@ -45,14 +45,10 @@ from protein_quest.alphafold.fetch import AlphaFoldEntry, DownloadableFormat
 from protein_quest.alphafold.fetch import fetch_many as alphafold_fetch
 from protein_quest.emdb import fetch as emdb_fetch
 from protein_quest.go import search_gene_ontology_term
+from protein_quest.io import convert_to_cif_file, glob_structure_files
 from protein_quest.pdbe.fetch import fetch as pdbe_fetch
-from protein_quest.pdbe.io import (
-    convert_to_cif_file,
-    glob_structure_files,
-    nr_residues_in_chain,
-    write_single_chain_pdb_file,
-)
 from protein_quest.ss import filter_file_on_secondary_structure
+from protein_quest.structure import nr_residues_in_chain, write_single_chain_structure_file
 from protein_quest.taxonomy import search_taxon
 from protein_quest.uniprot import (
     PdbResult,
@@ -128,7 +124,7 @@ def extract_single_chain_from_structure(
     Returns:
         Path to the output structure (mmCIF or pdb) file
     """
-    return write_single_chain_pdb_file(input_file, chain2keep, output_dir, out_chain)
+    return write_single_chain_structure_file(input_file, chain2keep, output_dir, out_chain)
 
 
 @mcp.tool
