+++ tests/conftest.py
@@ -11,3 +11,8 @@ def sample_cif() -> Path:
     """
     return Path(__file__).parent / "fixtures" / "3JRS_B2A.cif.gz"
 
+
+@pytest.fixture
+def sample2_cif() -> Path:
+    """Downloaded from https://www.rcsb.org/structure/2Y29"""
+    return Path(__file__).parent / "fixtures" / "2Y29.cif.gz"
